requirements.txt

Flask==1.1.1
Flask-Babel==0.12.2
Flask-Compress==1.4.0
Flask-Cors==3.0.6
Flask-RESTful==0.3.6

Keras==2.2.4

Markdown==3.1.1
nltk

numpy==1.17.2

pandas
pickleshare==0.7.5
Pillow==5.1.0

python-crfsuite==0.9.6

regex==2019.6.8

requests==2.22.0

scikit-learn==0.21.3

sklearn-crfsuite==0.3.6

Werkzeug==0.15.5

